+++ tests/test_schema.py
@@ -32,7 +32,7 @@ def test_mlm_schema(
     ["item_raster_bands.json"],
     indirect=True,
 )
-def test_mlm_no_undefined_prefixed_field(
+def test_mlm_no_undefined_prefixed_field_item_properties(
     mlm_validator: STACValidator,
     mlm_example: Dict[str, JSON],
 ) -> None:
@@ -40,14 +40,60 @@ def test_mlm_no_undefined_prefixed_field(
     mlm_item = pystac.Item.from_dict(mlm_data)
     pystac.validation.validate(mlm_item, validator=mlm_validator)  # ensure original is valid
 
+    # undefined property anywhere in the schema
+    mlm_data = copy.deepcopy(mlm_example)
     mlm_data["properties"]["mlm:unknown"] = "random"  # type: ignore
     with pytest.raises(pystac.errors.STACValidationError) as exc:
         mlm_item = pystac.Item.from_dict(mlm_data)
         pystac.validation.validate(mlm_item, validator=mlm_validator)
     assert all(
-        field in str(exc.value.source)
-        for field in ["mlm:unknown", "^(?!mlm:)"]
+        info in str(exc.value.source)
+        for info in ["mlm:unknown", "Unevaluated properties are not allowed"]
     )
+
+    # defined property only allowed at the Asset level
+    mlm_data = copy.deepcopy(mlm_example)
+    mlm_data["properties"]["mlm:artifact_type"] = "torch.save"
+    with pytest.raises(pystac.errors.STACValidationError) as exc:
+        mlm_item = pystac.Item.from_dict(mlm_data)
+        pystac.validation.validate(mlm_item, validator=mlm_validator)
+    assert all(
+        field in str(exc.value.source)
+        for field in ["mlm:artifact_type", "Unevaluated properties are not allowed"]
+    )
+
+
+@pytest.mark.parametrize(
+    "mlm_example",
+    ["item_raster_bands.json"],
+    indirect=True,
+)
+def test_mlm_no_undefined_prefixed_field_asset_properties(
+    mlm_validator: STACValidator,
+    mlm_example: Dict[str, JSON],
+) -> None:
+    mlm_data = copy.deepcopy(mlm_example)
+    mlm_item = pystac.Item.from_dict(mlm_data)
+    pystac.validation.validate(mlm_item, validator=mlm_validator)  # ensure original is valid
+
+    assert mlm_data["assets"]["weights"]
+    mlm_data["assets"]["weights"]["mlm:unknown"] = "random"  # type: ignore
+    with pytest.raises(pystac.errors.STACValidationError) as exc:
+        mlm_item = pystac.Item.from_dict(mlm_data)
+        pystac.validation.validate(mlm_item, validator=mlm_validator)
+    assert len(exc.value.source) == 1
+    schema_error = exc.value.source[0]
+    assert "mlm:unknown" in schema_error.instance
+    assert schema_error.schema["description"] in [
+        "Fields that apply only within an Asset.",
+        "Schema to validate the MLM fields permitted only under Assets properties."
+    ]
+
+    # defined property allowed both at the Item at the Asset level
+    mlm_data = copy.deepcopy(mlm_example)
+    mlm_data["assets"]["weights"]["mlm:accelerator"] = "cuda"
+    mlm_item = pystac.Item.from_dict(mlm_data)
+    pystac.validation.validate(mlm_item, validator=mlm_validator)
 
 
 @pytest.mark.parametrize(
@@ -222,6 +268,32 @@ def test_mlm_at_least_one_asset_model(
             "At least one Asset must provide the model definition indicated by the 'mlm:model' role.",
             "Used to check the artifact type property that is required by a Model Asset annotated by 'mlm:model' role."
         ]
+
+
+@pytest.mark.parametrize(
+    "mlm_example",
+    ["item_basic.json"],
+    indirect=True,
+)
+def test_mlm_asset_artifact_type_checked(
+    mlm_validator: STACValidator,
+    mlm_example: Dict[str, JSON],
+) -> None:
+    mlm_data = copy.deepcopy(mlm_example)
+    mlm_item = pystac.Item.from_dict(mlm_data)
+    pystac.validation.validate(mlm_item, validator=mlm_validator)  # self-check valid beforehand
+
+    mlm_data["assets"]["model"]["mlm:artifact_type"] = 1234
+    mlm_item = pystac.Item.from_dict(mlm_data)
+    with pytest.raises(pystac.errors.STACValidationError) as exc:
+        pystac.validation.validate(mlm_item, validator=mlm_validator)
+    assert "1234 is not of type 'string'" in str(exc.value.source)
+
+    mlm_data["assets"]["model"]["mlm:artifact_type"] = ""
+    mlm_item = pystac.Item.from_dict(mlm_data)
+    with pytest.raises(pystac.errors.STACValidationError) as exc:
+        pystac.validation.validate(mlm_item, validator=mlm_validator)
+    assert "should be non-empty" in str(exc.value.source)
 
 
 def test_model_metadata_to_dict(eurosat_resnet):